backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "DB app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "asgi-lifespan>=2.1.0",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "uvicorn>=0.34.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
